Editing OgeroRadiusCDR DB structure
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C49618

+++ Mediation/Code/Mediation.CDR/Mediation_WHS/User Defined Types/OgeroRadiusCDRType.sql
@@ -38,8 +38,8 @@
     [AcctSessionId]          VARCHAR (50) NULL,
     [AcctAuthentic]          INT          NULL,
     [AcctSessionTime]        VARCHAR (50) NULL,
-    [AcctInputPackets]       INT          NULL,
-    [AcctOutputPackets]      INT          NULL,
+    [AcctInputPackets]       BIGINT       NULL,
+    [AcctOutputPackets]      BIGINT       NULL,
     [AcctTerminateCause]     INT          NULL,
     [AcctMultiSsnID]         INT          NULL,
     [AcctLinkCount]          INT          NULL,
@@ -61,6 +61,8 @@
     [MSMPPEEncryptionTypes]  INT          NULL,
     [MSMPPEEncryptionPolicy] INT          NULL,
     [FileName]               VARCHAR (50) NULL);
+
+
